Extract Vedeu::Log subclasses to separate files.
