+++ lib/vedeu/all.rb
@@ -2,6 +2,8 @@ require 'vedeu/version'
 require 'vedeu/runtime/launcher'
 require 'vedeu/runtime/bootstrap'
 require 'vedeu/exceptions'
+require 'vedeu/log/lockless_log_device'
+require 'vedeu/log/mono_logger'
 require 'vedeu/log/log'
 require 'vedeu/log/debug'
 require 'vedeu/log/timer'
