+++ lib/vedeu/log/log.rb
@@ -1,81 +1,5 @@
 module Vedeu
 
-  # Allows the creation of a lock-less log device.
-  #
-  class MonoLogger < Logger
-
-    # Create a trappable Logger instance.
-    #
-    # @param logdev [String|IO] The filename (String) or IO object (typically
-    #   STDOUT, STDERR or an open file).
-    #
-    def initialize(logdev)
-      @level             = DEBUG
-      @default_formatter = Formatter.new
-      @formatter         = nil
-      @logdev            = LocklessLogDevice.new(logdev) if logdev
-    end
-
-    # Ensures we can always write to the log file by creating a lock-less
-    # log device.
-    #
-    class LocklessLogDevice < LogDevice
-
-      # Returns a new instance of Vedeu::LocklessLogDevice.
-      #
-      # @param log [void]
-      # @return [void]
-      def initialize(log = nil)
-        @dev      = nil
-        @filename = nil
-
-        if log.respond_to?(:write) && log.respond_to?(:close)
-          @dev = log
-
-        else
-          @dev      = open_logfile(log)
-          @dev.sync = true
-          @filename = log
-
-        end
-      end
-
-      # @param message [String]
-      # @return [void]
-      def write(message)
-        @dev.write(message)
-
-      rescue StandardError => exception
-        warn("log writing failed. #{exception}")
-      end
-
-      # @return [void]
-      def close
-        @dev.close
-      rescue
-        nil
-      end
-
-      private
-
-      # @param filename [String]
-      # @return [void]
-      def open_logfile(filename)
-        if FileTest.exist?(filename)
-          open(filename, (File::WRONLY | File::APPEND))
-
-        else
-          logdev = open(filename, (File::WRONLY | File::APPEND | File::CREAT))
-          logdev.sync = true
-          logdev
-
-        end
-      end
-
-    end # LocklessLogDevice
-
-  end # MonoLogger
-
   # Provides the ability to log anything to the Vedeu log file.
   #
   class Log
