+++ test/lib/vedeu/log/lockless_log_device_test.rb
@@ -1,0 +1,17 @@
+require 'test_helper'
+
+module Vedeu
+
+  describe LocklessLogDevice do
+
+    # let(:described) { Vedeu::LocklessLogDevice }
+    # let(:instance)  { described.new }
+    # # let(:log)       {}
+
+    # describe '#initialize' do
+    #   it { instance.must_be_instance_of(described) }
+    # end
+
+  end # LocklessLogDevice
+
+end # Vedeu
